refactor(in-app-agent): deepen background interfaces
