+++ packages/shared/src/in-app-agent/server/index.ts
@@ -3,6 +3,8 @@
 export * from "./access";
 export * from "./agent";
 export * from "./tools";
+export * from "./mcpPolicy";
+export * from "./toolResults";
 export * from "./human-in-the-loop";
 export * from "./instrumentation";
 export * from "./eventCompaction";
