+++ packages/shared/src/in-app-agent/server/instrumentation.ts
@@ -19,7 +19,7 @@ import {
   normalizeToolOutput,
   parseJsonOrString,
 } from "./toolErrors";
-import type { InAppAgentUserAccess } from "./tools";
+import type { InAppAgentUserAccess } from "./mcpPolicy";
 import { assertUnreachable } from "../../utils/typeChecks";
 
 export type InAppAgentTracingConfig = {
