+++ web/src/features/in-app-agent/README.md
@@ -25,8 +25,11 @@ Closing the drawer detaches observation without cancelling the worker run.
 
 - `schema.ts`: runtime-neutral AG-UI schemas and types shared by browser, server, persistence, replay, and rendering, including Langfuse-owned human-in-the-loop wire contracts.
 - `server/agent.ts`: Mastra/Bedrock/MCP runtime setup, custom tool wiring, human-in-the-loop approval gates, AG-UI event normalization, and cleanup.
-- `server/human-in-the-loop.ts`: interrupt parsing and worker continuation
+- `interrupts.ts`: browser-safe parsing of the durable approval interrupt
+- `server/human-in-the-loop.ts`: worker continuation handling
   compatibility.
+- `server/mcpPolicy.ts`: MCP names, authorization, and approval policy
+- `server/toolResults.ts`: persisted MCP result redaction and sandbox-file contracts
 - `server/tools.ts`: custom agent tools with strict schemas and scoped, user-visible behavior.
 - `server/persistence.ts`: conversations, events, canonical accumulation, and replay. Knows nothing about rendering.
 - `server/runLifecycle.ts`: durable run creation, claiming, cancellation, terminal transitions, active-run locking, and reconciliation.
@@ -99,7 +102,10 @@ flowchart TB
 
 ## Run Lifecycles
 
-1. The provider starts the durable run through `startRun`.
+1. The provider submits one user message plus browser context to the session.
+   The session owns optimistic insertion and constructs the internal AG-UI run
+   input before starting the durable run through `startRun`; the server remains
+   the authoritative sanitizer before persisting the request.
 2. `BackgroundExecutionSessionController` installs the persisted canonical
    messages, the display state, and the cursor before attaching the watch
    stream. The seed is never projected or pruned; see `ARCHITECTURE.md`.
@@ -210,13 +216,14 @@ MCP registry behavior:
 - In-app-agent keys can call read-only tools directly when the tool has `readOnlyHint: true`.
 - In-app-agent keys need a valid tool override to call named non-read-only Langfuse MCP tools.
 
-RBAC is the first gate for Langfuse MCP tools. Before a tool is exposed to the model, `server/tools.ts` checks the signed-in user's `projectRole` and `isAdmin` against the tool's required `ProjectScope` with `hasProjectAccess()`. That means the assistant never sees tools the user could not use manually in the product UI or APIs. Human approval is a second gate on top of RBAC for tools classified as `"approval"`: approval can allow one execution of a tool the user already has access to, but it does not widen the user's project permissions.
+RBAC is the first gate for Langfuse MCP tools. Before a tool is exposed to the model, `server/mcpPolicy.ts` checks the signed-in user's `projectRole` and `isAdmin` against the tool's required `ProjectScope` with `hasProjectAccess()`. That means the assistant never sees tools the user could not use manually in the product UI or APIs. Human approval is a second gate on top of RBAC for tools classified as `"approval"`: approval can allow one execution of a tool the user already has access to, but it does not widen the user's project permissions.
 
-Human approval is separate from the MCP tool override. Shared `server/tools.ts` classifies every Langfuse MCP tool in `IN_APP_AGENT_LANGFUSE_MCP_TOOL_POLICIES`, using unprefixed MCP registry names and either `"auto"` or `"approval"`. The web in-app-agent server test imports the MCP registry's `McpToolName` contract and verifies both type equality and runtime registry equality, so adding a Langfuse MCP tool requires an explicit in-app agent approval classification without making MCP bootstrap depend on the in-app agent.
+Human approval is separate from the MCP tool override. Shared `server/mcpPolicy.ts` classifies every Langfuse MCP tool in `IN_APP_AGENT_LANGFUSE_MCP_TOOL_POLICIES`, using unprefixed MCP registry names and either `"auto"` or `"approval"`. The web in-app-agent server test imports the MCP registry's `McpToolName` contract and verifies both type equality and runtime registry equality, so adding a Langfuse MCP tool requires an explicit in-app agent approval classification without making MCP bootstrap depend on the in-app agent.
 
-`server/tools.ts` turns that map into an `InAppAgentToolPolicy` with an `available` set for tools the model may see and an `autoApproved` set for tools that run without a prompt. `createInAppAgentToolPolicy()` intersects stored grants with current availability, so an unknown, deleted, reclassified, or now-out-of-role tool silently drops out. Local tools such as `IN_APP_AGENT_REDIRECT_TOOL_NAME` and the sandbox tools are always auto-approved; docs MCP tools are auto-approved by the `langfuseDocs_` prefix. `server/agent.ts` marks every other tool with Mastra `requireApproval: true`.
+`server/mcpPolicy.ts` turns that map into an `InAppAgentToolPolicy` with an `available` set for tools the model may see and an `autoApproved` set for tools that run without a prompt. `createInAppAgentToolPolicy()` intersects stored grants with current availability, so an unknown, deleted, reclassified, or now-out-of-role tool silently drops out. Local tools such as `IN_APP_AGENT_REDIRECT_TOOL_NAME` and the sandbox tools are always auto-approved; docs MCP tools are auto-approved by the `langfuseDocs_` prefix. `server/agent.ts` marks every other tool with Mastra `requireApproval: true`.
 
 Conversation-scoped grants live in `InAppAgentConversation.alwaysAllowedTools`, holding server-prefixed names such as `langfuse_createDashboardWidget`. The prefix is part of the identity, so a stored grant cannot be mistaken for a same-named tool on another MCP surface; unprefixed entries are rejected when the policy is rebuilt. Grants are written inside the same locked transaction as the approval decision in `server/runLifecycle.ts`. A one-off `Approve` writes nothing there; only `Always approve` does. The worker rebuilds the policy from the column on every run rather than trusting enqueue-time authorization. Mastra emits an interrupt, the browser asks the user, and the router records the decision for a durable worker continuation. `server/human-in-the-loop.ts` adapts Mastra's runtime interrupt payload into the Langfuse-owned `tool_approval_request` contract from `schema.ts`; the browser stores and forwards only that runtime-neutral shape.
+The browser-safe `interrupts.ts` parser adapts Mastra's runtime payload into the Langfuse-owned `tool_approval_request` contract from `schema.ts`; browser, web server, and worker all consume that same parser.
 
 The `InAppAgentPendingToolApproval` table is not used by background execution.
 It remains temporarily so existing rows and Prisma relations stay valid.
