--- web/src/features/in-app-agent/components/index.ts
@@ -1,2 +1,0 @@
-export { ControlledInAppAgentWindow } from "./ControlledInAppAgentWindow";
-export { InAppAiAgentProvider, useInAppAiAgent } from "./InAppAiAgentProvider";
