+++ web/src/features/in-app-agent/components/utils/utils.ts
@@ -788,7 +788,7 @@ function getRedirectActionFromToolResult(
   }
 }
 
-export function extractLangfuseDocsSources(
+function extractLangfuseDocsSources(
   tools: readonly InAppAgentToolCallContent[],
 ): InAppAgentMessageSource[] {
   return mergeSources(
