+++ web/src/features/in-app-agent/context.ts
@@ -1,8 +1,4 @@
 import type { AgUiRunAgentInput } from "@langfuse/shared/in-app-agent";
-import {
-  isInAppAgentQuickActionContext,
-  type InAppAgentQuickActionAttribution,
-} from "@/src/features/in-app-agent/quickActions";
 import { getInAppAgentProjectRoute } from "@/src/features/in-app-agent/routeContext";
 import type { FilterState } from "@langfuse/shared";
 
@@ -30,17 +26,12 @@ export type InAppAgentScreenContextDescription =
   | { type: "datasets-list" };
 
 const CURRENT_URL_CONTEXT_DESCRIPTION = "current_url";
-const QUICK_ACTION_KEY_CONTEXT_DESCRIPTION = "quick_action_key";
-const QUICK_ACTION_CATEGORY_CONTEXT_DESCRIPTION = "quick_action_category";
-const MESSAGE_ENTRY_POINT_CONTEXT_DESCRIPTION = "message_entry_point";
 const MAX_SCREEN_CONTEXT_SEARCH_PARAMS = 30;
 const MAX_CONTEXT_KEY_LENGTH = 80;
 const MAX_CONTEXT_VALUE_LENGTH = 500;
 const MAX_SCREEN_CONTEXT_PATH_LENGTH = 500;
 const MAX_SCREEN_CONTEXT_HASH_LENGTH = 200;
 const MAX_SCREEN_CONTEXT_JSON_LENGTH = 4_000;
-const MAX_QUICK_ACTION_KEY_LENGTH = 80;
-const QUICK_ACTION_KEY_PATTERN = /^[a-z0-9]+(?:-[a-z0-9]+)*$/;
 const USER_CONTEXT_DESCRIPTIONS = new Set([
   "user_name",
   "current_timezone",
@@ -207,99 +198,6 @@ export function sanitizeInAppAgentContext(
   sanitizedContext.push(...sanitizeUserContext(context));
 
   return sanitizedContext;
-}
-
-export function createInAppAgentQuickActionAttributionContext(
-  attribution: InAppAgentQuickActionAttribution,
-): InAppAgentContext {
-  return [
-    {
-      description: QUICK_ACTION_KEY_CONTEXT_DESCRIPTION,
-      value: attribution.key,
-    },
-    {
-      description: QUICK_ACTION_CATEGORY_CONTEXT_DESCRIPTION,
-      value: attribution.category,
-    },
-  ];
-}
-
-// Attribution is telemetry only: it is validated by shape here and read for
-// trace metadata, but never forwarded into the model-visible sanitized context.
-export function getInAppAgentQuickActionAttribution(
-  context: InAppAgentContext,
-): InAppAgentQuickActionAttribution | undefined {
-  const quickActionKey = context
-    .find((item) => item.description === QUICK_ACTION_KEY_CONTEXT_DESCRIPTION)
-    ?.value.trim();
-  const quickActionCategory = context
-    .find(
-      (item) => item.description === QUICK_ACTION_CATEGORY_CONTEXT_DESCRIPTION,
-    )
-    ?.value.trim();
-
-  if (
-    !quickActionKey ||
-    quickActionKey.length > MAX_QUICK_ACTION_KEY_LENGTH ||
-    !QUICK_ACTION_KEY_PATTERN.test(quickActionKey) ||
-    !quickActionCategory ||
-    !isInAppAgentQuickActionContext(quickActionCategory)
-  ) {
-    return undefined;
-  }
-
-  return { key: quickActionKey, category: quickActionCategory };
-}
-
-export function getInAppAgentQuickActionTraceMetadata(
-  context: InAppAgentContext,
-): Record<string, string> {
-  const attribution = getInAppAgentQuickActionAttribution(context);
-
-  return attribution
-    ? {
-        quick_action_key: attribution.key,
-        quick_action_category: attribution.category,
-      }
-    : {};
-}
-
-export const IN_APP_AGENT_MESSAGE_ENTRY_POINTS = [
-  "chat",
-  "add-widget-modal",
-] as const;
-
-export type InAppAgentMessageEntryPoint =
-  (typeof IN_APP_AGENT_MESSAGE_ENTRY_POINTS)[number];
-
-export function createInAppAgentMessageEntryPointContext(
-  entryPoint: InAppAgentMessageEntryPoint,
-): InAppAgentContext {
-  return [
-    {
-      description: MESSAGE_ENTRY_POINT_CONTEXT_DESCRIPTION,
-      value: entryPoint,
-    },
-  ];
-}
-
-// Telemetry only, like quick-action attribution: read for trace metadata but
-// never forwarded into the model-visible sanitized context.
-export function getInAppAgentMessageEntryPointTraceMetadata(
-  context: InAppAgentContext,
-): Record<string, string> {
-  const entryPoint = context
-    .find(
-      (item) => item.description === MESSAGE_ENTRY_POINT_CONTEXT_DESCRIPTION,
-    )
-    ?.value.trim();
-
-  return entryPoint &&
-    (IN_APP_AGENT_MESSAGE_ENTRY_POINTS as readonly string[]).includes(
-      entryPoint,
-    )
-    ? { message_entry_point: entryPoint }
-    : {};
 }
 
 function sanitizeUserContext(context: InAppAgentContext): InAppAgentContext {
